Initial commit of current Eclipse project
Heavily modified code of optaplannerhello to test Hibernate,/MySQL
connectivity with the LittleSprouts database, xStream, Optaplanner core,
and built using Maven. All POJOs are created for use with the database.
Testing session with database created to test functionality of database
transactions. Will need to add a .gitignore if Eclipse workspace is
created using this project