MVC version of project and gitignore
Changed project to MVC for easier web deployment and development. Import
project with the .pom file. To run, right click project, run as, and
either create or select Tomcat 7 server. Change the user login
credentials for the database in the jpaContext.xml file (its in the last
defined bean). May need Maven installed on PC and Spring Tools Suite
plugin installed in Eclipse to build project.